fix(game): holdrand is platform-width c_ulong, not u32 — t2_wedge NPC-type pick parity
Raven's LCG state is `static unsigned long holdrand` (q_math.c:1432) —
32-bit on the retail i686 ship but 64-bit on LP64 referee/native builds,
where `result = holdrand >> 17` pulls in high bits a u32 model truncates.
The low 32 bits of the stream agree on both models, so the divergence only
shows where a draw's RESULT differs: on t2_wedge every Q_irand NPC-type
pick (SP_NPC_Stormtrooper StormTrooper/StormTrooper2, SP_NPC_ImpWorker
3-way) chose differently, registering extra *$st2/*$worker3 sound sets
and shifting all later CS_SOUNDS indices (the csSounds_Std diffs on ~60
NPC spawners).

Ruling 2026-07-09 (docs/handoffs/jampgame-fork-discovery.md): holdrand is
c_ulong — platform-faithful on every target. Reverses the 2026-07 u32
normalization: rng.rs u32 state + assert, and tools/jampgame-oracle's
'unsigned long -> unsigned int' rewrite (run.sh) now extracts the LCG at
native width; qmath.txt rng-section goldens regenerated (3083 lines) and
reproduced by cargo test. pmove parity tripwires truncate to u32 at the
call site (their fixtures draw nothing).

Referee: t2_wedge frame-1 csSounds divergence gone (remaining: apos
sub-ULP on ents 60/156, separate issue); both ffa1 corpora PASS 4000.

Source: oracle/codemp/game/q_math.c:1425-1474

+++ crates/mp/game/tests/pmove_parity.rs
@@ -944,7 +944,7 @@ fn run_scenario(name: &str) -> String {
                 // pm.ps / baseEnt must point at the live ps / arena for the dump.
                 pm.ps = &mut ps as *mut playerState_t;
                 pm.baseEnt = arena.as_mut_ptr() as *mut _;
-                let rng = bg.rng.holdrand();
+                let rng = bg.rng.holdrand() as u32; // 32-bit tripwire (fixtures draw nothing)
                 dump_step(&mut o, step, &pm, &ps, traps.trace_count.get(), rng);
                 step += 1;
             }
@@ -971,7 +971,7 @@ fn run_scenario(name: &str) -> String {
                     cb.legs_mirror = ps.legsAnim;
                     cb.torso_mirror = ps.torsoAnim;
 
-                    let rng = bg.rng.holdrand();
+                    let rng = bg.rng.holdrand() as u32; // 32-bit tripwire (fixtures draw nothing)
                     dump_step(&mut o, step, &pm, &ps, traps.trace_count.get(), rng);
                     step += 1;
                 }
